Fixed MAX30105 driver implementation

+++ src/i2cmaster.h
@@ -1,6 +1,7 @@
 #ifndef _include_i2cmaster_h_
 #define _include_i2cmaster_h_
 
+#include "blackbox.h"
 #include <cstdint>
 
 class i2c_master
@@ -10,10 +11,10 @@ public:
 	i2c_master(int bus);
 	void set_slave_address(uint8_t id);
 	void write_bytes(void *data, uint8_t length);
-	size_t read_bytes(void *data, uint8_t length);
+	uint8_t read_bytes(void *data, uint8_t length);
 	void write_register(uint8_t reg, uint8_t value);
 	void write_register16(uint8_t reg, uint16_t value);
-
+	
 	template<typename T>
 	T read_register(uint8_t reg)
 	{
